Learning the basics of CATIA
Using a flange assemble for practice, I learnt most of the basics commands in CATIA. Coming from Solidworks, the concepts were relatively easy, however, it was difficult to get the hang of it since it is not as easy to use as Solidworks. Maybe with practice, I will be able to use CATIA as easily as I do Solidworks. In this exercise, I learnt the following:
1. CATIA Sketcher
2. Basic commands in CATIA
3. Multiple views
4. Constraints and snapping in assemble
5. Creating a drawing from the final assembly
6. Projecting different views (top, front, side, and orthographic, as wel as cut-section)
7. Dimensioning in drawing
8. Exporting the drawing as a PDF.

You may see the final drawing in the "Flange-Assembly-Drawing.pdf" file and the final assembly in the "Flange-Assembly.CATProduct" file.